Message cleanup
git-svn-id: file:///home/jj/hercules.svn/trunk@6980 956126f8-22a0-4046-8f4a-272fa8102e63

+++ cardpch.c
@@ -62,7 +62,7 @@ int     i;                              /* Array subscript           */
 
     if (strlen(argv[0]) >= sizeof(dev->filename))
     {
-        WRMSG (HHC01201, "E", SSID_TO_LCSS(dev->ssid), dev->devnum, argv[0], sizeof(dev->filename) - 1);
+        WRMSG (HHC01201, "E", SSID_TO_LCSS(dev->ssid), dev->devnum, argv[0], (int)sizeof(dev->filename) - 1);
         return -1;
     }
